added connection status end point and election error handling
